Add fathom-test tool. Fix #90.
* Move Python commands down a level into a `commands` folder.
* Factor out common bits I need from fathom-train.
