+++ cli/setup.py
@@ -14,10 +14,11 @@ setup(
     url='https://mozilla.github.io/fathom/',
     install_requires=['click>=7.0,<8.0', 'tensorboardX>=1.6,<2.0', 'torch>=1.0,<2.0'],
     entry_points={'console_scripts': [
-        'fathom-train = fathom_web.train:main',
-        'fathom-unzip = fathom_web.unzip:main',
-        'fathom-pick = fathom_web.pick:main',
-        'fathom-list = fathom_web.list:main'
+        'fathom-test = fathom_web.commands.test:main',
+        'fathom-train = fathom_web.commands.train:main',
+        'fathom-unzip = fathom_web.commands.unzip:main',
+        'fathom-pick = fathom_web.commands.pick:main',
+        'fathom-list = fathom_web.commands.list:main'
     ]},
     classifiers=[
         'Intended Audience :: Developers',
